Remove Search contracts.
