+++ mastodon4j-rx/src/main/java/com/sys1yagi/mastodon4j/rx/RxPublic.kt
@@ -2,6 +2,7 @@ package com.sys1yagi.mastodon4j.rx
 
 import com.sys1yagi.mastodon4j.MastodonClient
 import com.sys1yagi.mastodon4j.api.entity.Instance
+import com.sys1yagi.mastodon4j.api.entity.Results
 import com.sys1yagi.mastodon4j.api.method.Public
 import com.sys1yagi.mastodon4j.rx.extensions.onErrorIfNotDisposed
 import io.reactivex.Single
@@ -16,6 +17,17 @@ class RxPublic(client: MastodonClient) {
                 it.onSuccess(instance)
             } catch(throwable: Throwable) {
                 it.onErrorIfNotDisposed(throwable)
+            }
+        }
+    }
+
+    fun getSearch(query: String, resolve: Boolean = true): Single<Results> {
+        return Single.create {
+            try {
+                val results = publicMethod.getSearch(query, resolve)
+                it.onSuccess(results)
+            } catch (e: Throwable) {
+                it.onError(e)
             }
         }
     }
